+++ R/elmt-methods.R
@@ -6,9 +6,9 @@ setMethod("elmt", "EL", function(object,
                                  alpha = 0.05,
                                  control = el_control()) {
   stopifnot(
-    "`object` has no `data`; fit the model with `model` = TRUE." =
-      (length(getDataMatrix(object)) > 1L),
-    "invalid `control` specified." = (is(control, "ControlEL"))
+    "`object` has no `data`. Fit the model with `keep_data == TRUE`." =
+      (!is.null(getDataMatrix(object))),
+    "Invalid `control` specified." = (is(control, "ControlEL"))
   )
   h <- validate_hypotheses(rhs, lhs, object@npar)
   alpha <- validate_alpha(alpha)