+++ R/print-methods.R
@@ -3,11 +3,6 @@
 #'   passed to [format()].
 #' @param ... Further arguments passed to or from other methods.
 #' @rdname EL-class
-#' @srrstats {RE3.0} Convergence status is printed at the bottom with other
-#'   information. It is well known that empirical likelihood methods can often
-#'   fail to converge due to the convex hull constraint. We expect that the
-#'   users of the package are also aware of this issue. Therefore, an explicit
-#'   warning message via `warning()` is issued.
 #' @srrstats {RE4.17} `print` method is applicable to an `EL` object. The test
 #'   results are displayed, including the type of test, test statistic,
 #'   coefficients, p-value, and convergence status.